refactor: change location of package sliders home.
