+++ app/src/main/java/com/rubio/movstore/ui/home/sliders/SliderTwoFragment.kt
@@ -1,22 +1,19 @@
-package com.rubio.movstore.ui.movcatalogue.sliders
+package com.rubio.movstore.ui.home.sliders
 
 import android.os.Bundle
+import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.databinding.DataBindingUtil
-import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
 import com.rubio.movstore.R
-import com.rubio.movstore.databinding.FragmentSliderThreeBinding
+import com.rubio.movstore.databinding.FragmentSliderTwoBinding
 import com.rubio.movstore.ui.home.viewmodel.HomeViewModel
 
-
-class SliderThreeFragment : Fragment() {
-
+class SliderTwoFragment : Fragment() {
     private val homeViewModel: HomeViewModel by activityViewModels()
-    private lateinit var binding: FragmentSliderThreeBinding
-
+    private lateinit var binding: FragmentSliderTwoBinding
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
@@ -27,9 +24,9 @@ class SliderThreeFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        binding =
-            DataBindingUtil.inflate(inflater, R.layout.fragment_slider_three, container, false)
+        binding = DataBindingUtil.inflate(inflater, R.layout.fragment_slider_two, container, false)
         binding.sliderModel = homeViewModel
         return binding.root
     }
+
 }
