+++ app/src/main/res/layout/fragment_slider_one.xml
@@ -14,7 +14,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:background="#CDDC39"
-        tools:context=".ui.movcatalogue.sliders.SliderOneFragment">
+        tools:context=".ui.home.sliders.SliderOneFragment">
 
         <ImageView
             android:id="@+id/SliderImg1"
