+++ app/src/main/res/layout/fragment_slider_two.xml
@@ -14,7 +14,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:background="#673AB7"
-        tools:context=".ui.movcatalogue.sliders.SliderTwoFragment">
+        tools:context=".ui.home.sliders.SliderTwoFragment">
 
         <ImageView
             android:id="@+id/SliderImg2"
